refactor : MainDomain 리팩토링
