+++ src/js/index.js
@@ -1,3 +1,2 @@
 import MainDomain from "./Domain/MainDomain.js";
-import ContestantView from "./ContestantView.js";
-new MainDomain(new ContestantView());
+new MainDomain();
